feature: create Icon component
