+++ src/components/IconContainer.vue
@@ -1,0 +1,46 @@
+<template>
+    <div class="icon">
+        <img :src="selectedIcon">
+    </div>
+</template>
+
+<script setup>
+import vue from "../assets/vue.svg"
+import darkMode from "../assets/icons/dark-mode.svg"
+import world from "../assets/icons/world.svg"
+import notification from "../assets/icons/notifications.svg"
+import { onMounted, ref } from "vue"
+
+const props = defineProps({
+    iconName: {
+        type: String,
+        required: true
+    }
+})
+
+let icons = {
+    vue,
+    darkMode,
+    world,
+    notification
+}
+
+let selectedIcon = ref("")
+
+onMounted(() => {
+    selectedIcon.value = getSVG()
+})
+
+function getSVG() {
+    return icons[props.iconName]
+}
+
+
+</script>
+
+<style scoped>
+.icon {
+    width: 35px;
+    height: 35px;
+}
+</style>
